requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
Flask==2.0.1
Flask-WTF==0.15.1
greenlet==1.1.1
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.19.5
pandas==1.1.5
psycopg2-binary==2.9.1
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.22
urllib3==1.26.6
Werkzeug==2.0.1
WTForms==2.3.3
